Added SAM measurement rate and altered how the close shield command works
The close servo command needed to override whatever position it was in. We would get issues where the cansat would think it is closed, but it is actually open. So the functionality of the close shield command no closes the shield for 3 seconds. It is meant to be pressed multiple times now.